[Cleanup] Small refactor to BigQueryViewSubDagCollector (Recidiviz/recidiviz-data#31899)
## Description of the change

A small refactor to the `BigQueryViewSubDagCollector` which is used by
the `load_views_to_sandbox` script. That class used to take both a list
of datasets and a list of addresses to filter the DAG down to. While
it's useful to be able to specify either a set of datasets or addresses
at the script args level, there's no need to have both these options on
the internal class when we can easily translate datasets to the
addresses of the views in those datasets.

Successfully ran a sandbox run with this command:
```
    python -m recidiviz.tools.load_views_to_sandbox \
        --sandbox_dataset_prefix ageiduschek manual \
        --view_ids_to_load sessions.location_type_sessions \
        --dataset_ids_to_load reference_views \
        --update_ancestors False \
        --update_descendants False
```

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

N/A - bug bash week cleanup

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [x] This pull request has a descriptive title and information useful
to a reviewer
- [x] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 4d43e94bd8e36c2eebf6616b9ac99e0a4dbf6f50