[Cleanup] Refactor DatasetReference types out of BigQueryClient interface (Recidiviz/recidiviz-data#31875)
## Description of the change

Every time I have to use a `BigQueryClient` method that takes in a
`DatasetReference` I am annoyed that I have to call
`dataset_ref_for_id()` to get a `DatasetReference` from the `dataset_id`
that I have available. This PR updates the `BigQueryClient` public
methods to only take in string `dataset_id` values.

In a follow-up PR, I would like to update all `BigQueryClient` methods
that take in a `dataset_id` and `table_id` pair to instead take in a
`BigQueryAddress` since that is the type we're using in most places and
it's less error-prone!

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

N/A - bug bash week!

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [x] This pull request has a descriptive title and information useful
to a reviewer
- [x] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 6763bd6f318625838c19c8431414b1519e0d2efa