[Tools] Remove ability to load all downstream views from load_views_to_sandbox auto mode (Recidiviz/recidiviz-data#31940)
## Description of the change

Now that we have unittests that check that our full view graph compiles,
there's no real reason to force users of the `load_views_to_sandbox`
`auto` mode to load all views downstream of their changes.

This PR adds a new `--load_changed_views_only` flag that allows you to
load just views changed on your branch and requires that either this
flag or one of the `--load_up_to*` flags is used.

If we have a view dependency chain A -> B -> C -> D and there are
changes to A and C on the branch, script runs with the
`--load_changed_views_only` flag will also load B so that changes in A
are reflected in the sandbox version of C.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

N/A - bug bash week!

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [x] This pull request has a descriptive title and information useful
to a reviewer
- [x] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: c0d5e36322ac35171942d85630826b933aca90c0