[Justice Counts] update agency dashboard pull to include systems and api link (Recidiviz/recidiviz-data#31948)
## Description of the change

Updates the agency dashboard pull to include a system column and the
link to the new v2 agency dashboard API. I ran the script locally to
confirm that the columns were added correctly. You can see it
[here](https://docs.google.com/spreadsheets/d/1TcskbQat7a3OA7X-KRra74qBcXZMZVg5nOfdRFa9-64/edit?gid=296011674#gid=296011674)
in on the `7-31-2024` tab.

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 4b227326b69892d10239dc8e4ffdba1c5abca761