[Justice Counts]  fix is_v0_dashboard_enabled assignment and remove value field for dimensions for agency dashboard (Recidiviz/recidiviz-data#31988)
## Description of the change

This PR introduces two small fixes to the published data endpoints. 1)
makes sure that `is_v0_dashboard_enabled` is only in the response for
the v2 published datapoints. **This is currently breaking the agency
dashboards, so let's get this fix in quickly.** 2) Removes the "value"
key/value for dimensions in v2 responses. We are sending all historical
data so the "value" is always None.

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: d2435a4e875dd90a76db4622e8da1392a04703c7