remove compliant reporting views that are no longer used (Recidiviz/recidiviz-data#31941)
## Description of the change

https://github.com/Recidiviz/looker/pull/586 deleted the now-unneeded
compliant reporting dashboard, which allows us to delete the views it
references. This PR deletes those views and continues down the chain for
all views that are now no longer being used (except for
`us_tn_standards_due`, which is being kept as requested).

~This PR is blocked on https://github.com/Recidiviz/looker/pull/587,
which removes a reference to one of these in looker.~ <-- has been
merged now! I checked the looker codebase and did not find any other
references to views/tables being deleted in this PR.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes https://github.com/Recidiviz/recidiviz-dashboards/issues/5003
Closes https://github.com/Recidiviz/recidiviz-data/issues/19015
Closes https://github.com/Recidiviz/recidiviz-data/issues/17885

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 441624cbebcc14959a81445cc6f31bc46d2dc549