Uses liquid filtering to create better guardrails around metrics for multiple task types (Recidiviz/recidiviz-data#31504)
## Description of the change

This is a change to our LookML autogeneration code will allow us to get
aggregations of distinct active users across all task types; currently,
that cross-opportunity analysis doesn't work because the aggregation
function for the EventDistinctUnitCount metric is just AVG, so querying
for distinct active users across all task types only yields the average
MAUs across all the different task type, rather than an aggregation over
all users.

It does so by adding "liquid wrapping" around all references to the json
field filter so that the filter can only be used if the associated
dimension is selected. This means that if we only have a single json
field filter (e.g., task type), only one of the following two behaviors
is allowed:
1. No filter applied to `task_type_filter` --> metrics are calculated as
a single aggregate across all tasks
2. Filter applied to `task_type_filter` --> task_type must be included
in the final query, and all metrics will be disaggregated by task type

We do want to count task completions for distinct opportunities as
distinct event counts even if no filter is selected -- so this also adds
task_type as an event segmentation column for all configured workflows
metrics that are not monthly active users metrics.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes #XXXX

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [ ] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

---------

Co-authored-by: mayukas <mayuka@stanford.edu>
GitOrigin-RevId: 2f206c0683c9abd523fb2934b858f1bd7023974a