[Cleanup] Update the BigQueryClient interface to use BigQueryAddress where possible (Recidiviz/recidiviz-data#31970)
## Description of the change

As a follow-up to Recidiviz/recidiviz-data#31875, updates all `BigQueryClient` methods that take
both a `dataset_id` and a `table_id` to instead take a
`BigQueryAddress`. For all functions that take in a `BigQueryAddress`
(or just a `dataset_id` for that matter, the implied project for that
address is the project associated with the `BigQueryClientImpl`
instance.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

N/A - bug bash

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [x] This pull request has a descriptive title and information useful
to a reviewer
- [x] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 3e917d0571f428837cca20b6eff97b1cac410c9c