use `supervision_staff_attribute_sessions` in roster sync query (Recidiviz/recidiviz-data#31907)
## Description of the change

The roster sync query is using a pretty similar query to
`supervision_staff_attribute_sessions`, so replace it with that. This
also lets us get rid of some deduping logic, since
`supervision_staff_attribute_sessions` already does that.

Loaded to sandbox at `dana_ingested_roster_sessions`. compare_views
shows exactly one diff, and it's a user that is newly missing. This user
has an entry in state_staff but all their sessions ended years ago, so I
think not having them in the roster is correct (they also don't appear
in auth0).

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes #XXXX

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [ ] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 8b4f523593690894bb665a0a380f97fe35d5e7d5