Add denial_reasons in WORKFLOWS_PERSON_IMPACT_FUNNEL_STATUS_SESSION (Recidiviz/recidiviz-data#31639)
## Description of the change
> Pulled denial_reasons from workflows_views.clients_snooze_spans into
analyst_data. all_task_type_marked_ineligible_spans so we are able to
incorporate denial_reasons into analyst_data.
workflows_person_impact_funnel_status_sessions_materialized for all
spans of time where someone is marked ineligible. Previously, the query
used DISTINCT but it is not possible to use DISTINCT directly to an
array, so we are using UNNEST with a CROSS JOIN and then a GROUP BY to
have the unique observations and its denial_reasons

## Type of change

Type: Feature

## Related issues

Closes Recidiviz/recidiviz-data#31622

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 28c97d3ad99bbf46bf6e6ac486634c97c4be3fea