[SFTP][Airflow] Only normalize dashes to underscores in base directory of download path (Recidiviz/recidiviz-data#31801)
## Description of the change

Moves the `replace("-", "_")` in `build_download_path` so that rather
than applying to the full file path, it only applies to the base
directory component. The only filenames that include dashes among tables
received via SFTP are for `SUPERVISIONCONTACT` in AR, which is received
in chunks and needs to use dashes rather than underscores for the chunks
to be recognized as part of the same file. Otherwise, the only dashes
that occur in filepaths for SFTP tables occur within base directory
names, so download paths for other tables will be unaffected by this
change.

Using the following query, we verified that no other regions have `-` in
the `remote_file_name` so we do no expect this change to have any effect
on other states' SFTP.

```
select region_code, count(*) from direct_ingest_sftp_ingest_ready_file_metadata where remote_file_path similar to '%/[^/]*-[^/]*.(csv|txt)' group by region_code;
```

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [x] Potential security implications or infrastructural changes have
been considered, if relevant

---------

Co-authored-by: ethanoro <ethanoro17@gmail.com>
GitOrigin-RevId: 894da3679c53d0b1b358785d136a311a04728ac5