[Testing] Update and Document Defaults in Ingest Pipeline Tests (Recidiviz/recidiviz-data#31958)
## Description of the change

This PR came about from wanting to integrate the new raw data fixture
loading of the ingest view tests into our pipeline tests. I learned a
lot about our existing tests, made some updates, and felt like this was
a good stopping point. This PR does a few things.

- Moves the `DirectIngestRawFixtureLoader`
- We now pass in a region module to the `DirectIngestRawFixtureLoader`
to allow fake regions and template regions to use it.
- Moves the default upper bound dates from our state specific tests to
the `DirectIngestRawFixtureLoader` as it already has access to all of
the raw tables in the region.
- Updated region_code to state_code in relevant tests.
- The `BaseStateIngestPipelineTestCase` is now a
`BigQueryEmulatorTestCase`
- Cleaned up some defaults between `BaseStateIngestPipelineTestCase`,
`StateIngestPipelineTestCase`, and
`StateSpecificIngestPipelineIntegrationTestCase`. We were setting region
information in each class, but then also overriding that information
with mocks in different `setUps`
- Updated `StateIngestPipelineTestCase` to use US_DD for all tests, the
existing comments made it seem we want to use that region for dataflow
tests and it made for some consistency

I found this was a good stopping point because
`StateIngestPipelineTestCase` has a different set of assumptions for
fixtures than `StateSpecificIngestPipelineIntegrationTestCase` and our
ingest view tests! I noted our existing TODO in various places where
this was relevant.

The `StateIngestPipelineTestCase` is meant to test functionality of the
pipeline itself and uses a fake state, while
`StateSpecificIngestPipelineIntegrationTestCase` is meant to provide a
full run of states that we are ingesting.

So, future work will be in two areas:
1. General quality of life improvements on comparing expected and actual
output (what I really wanted to accomplish this week)
2. Standardizing fixture assumptions and loading between pipeline and
ingest view tests. I think ideally the metadata defaults across fixtures
should be the same across our tests. Then we can load everything to the
emulator with the same code!

GitOrigin-RevId: 0fe54990f394e94eda82a425bbc08c939895d5c0