[US_TN] Fixing edge cases with Unit movements after release (Recidiviz/recidiviz-data#31987)
## Description of the change

I am adding flags for when there is an open period beginning with a
movement type of unit_change and a flag for when the previous end
movement type ends in a code indicating it is a transition to
supervision or out of the system. We want to exclude the periods when
these flags are both true so we do not create falsely open periods
because of unit changes incorrectly time-stamped after these key out
movements

The ingest pipeline ran successfully.

ingest_view_results with fix: 7859099
ingest_view_results before fix: 7859108
difference of 9

Querying the current view with this filter results in 9
```
WHERE last_movement = 'release' and open_with_unit_change = 'yes'
```
Confirmed all 9 of these people's last movements in OffenderMovement are
a movement to supervision or discharged

I also confirmed this resolves the issue for the person surfaced in this
ticket Recidiviz/recidiviz-data#5939

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes Recidiviz/recidiviz-data#5939

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [x] This pull request has a descriptive title and information useful
to a reviewer
- [x] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: aa6ad19fc0f2e6bd55113c5d725c01da07523df0