[US_ND] Remap StateIncarcerationType and StateCustodialAuthority (Recidiviz/recidiviz-data#32096)
## Description of the change

This PR accomplishes two things:

1. Remaps custodial authority and incarceration type for `GENERAL`
incarceration periods within the CJ county jail facility so that they
are considered "included" in the state population and will be counted in
sessions view.

ND has started using county jails to help relieve prison overcrowding
and so there has been an increase in residents assigned to county jails
while serving state prison sentences. In order for these residents to be
considered eligible for opportunities like work release they need to be
classified as "included in state".

So, when `StateSpecializedPurposeForIncarceration = GENERAL` and
`Facility = "CJ"`, then `StateCustodialAuthority = STATE_PRISON` and
`StateIncarcerationType = COUNTY_JAIL`. This includes scenarios where
someone is on work release and incarcerated in a county jail.

2. Remaps custodial authority and incarceration type when `facility =
"NTAD" # (Not Admitted)` so that `StateIncarcerationType = OUT_OF_STATE`
and `StateCustodialAuthority = STATE_PRISON`

Along with that change, I also remapped `StateIncarcerationType` to be
`OUT_OF_STATE` when facility is in 'OOS', 'OS', or 'OUT', which were
assumed to refer to out of state assignments but were not already mapped
to that enum value.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes Recidiviz/recidiviz-data#31973
Closes Recidiviz/recidiviz-data#32122

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [ ] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 619c90e9fc1dd4cb5a7c5cd997341a37046e166a