Fix Insights/Workflows configuration pages in Admin Panel (Recidiviz/recidiviz-data#32160)
## Description of the change

[This PR](https://github.com/Recidiviz/recidiviz-data/pull/32058) caused
another issue ... I assumed that all databases had corresponding host
secrets in GCP, but apparently WFs and Insights do not, they only have
those secrets locally. After my change, the host is only needed if
using_unix_sockets=False (which is currently only the case when running
locally) -- I was just pulling it in all cases. So, we can refactor this
so we only grab the host if using_unix_sockets=False.

It probably can't hurt to add the host secrets to GCP, too.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes #XXXX

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [ ] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 784904bb230bd56acfccb1fc4a4a203d535ce0ed