[Justice Counts] Update commits_to_be_deployed script to support array and string tags (Recidiviz/recidiviz-data#32146)
## Description of the change

Update `commits_to_be_deployed` script to support array and string tags.
This is caused by differing versions of `gcloud` where one produces an
array of strings for tags, and the other a comma-separated string list.

Tested by using different versions of `gcloud` (version 487.0.0 and
472.0.0, respectively):

![Screenshot 2024-08-07 at 11 48
34 AM](https://github.com/user-attachments/assets/2475789e-fdbc-4810-b434-2d9461aa55e0)

![Screenshot 2024-08-07 at 11 42
36 AM](https://github.com/user-attachments/assets/9c687743-4ef0-461f-a945-90f52a3873de)

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes #XXXX

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [ ] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

---------

Co-authored-by: Mahmoud <mahmoud@Mahmouds-MacBook-Pro.local>
GitOrigin-RevId: 281c10353d1e2bf9c59f9756fa9d0f6ed8154524